Add CSS file, UI and functionality
Added UI, so user can play the game on the page. Added buttons for selection
and a select element of how many times to play. Added eventlisteners on the buttons
like adding player and cpuscores and displaying message on the page every round
played. Created a stylesheet to center elements on the page.

[Ticket: X]